model_features/requirements.txt
pandas
scikit-learn
psycopg2-binary
boto3
pyarrow
fastparquet
python-dotenv


